Extract LoginButton Component
