+++ client/package.json
@@ -14,7 +14,6 @@
     "react": "^16.1.0",
     "react-apollo": "^2.0.4",
     "react-dom": "^16.1.0",
-    "react-google-login": "^3.0.4",
     "react-moment": "^0.6.8",
     "react-router-dom": "^4.2.2",
     "react-scripts-ts": "2.8.0",
